// src/recipe/thirdparty/types.ts
export interface UserInfo { id: string; email?: { id: string; isVerified: boolean } }

export interface TypeProviderGetResponse {
  accessTokenAPI: {
    url: string
    params: { [key: string]: string } // Will be merged with our object
  }
  authorisationRedirect: {
    url: string
    params: { [key: string]: string | ((request: any) => string) }
  }
  getProfileInfo: (authCodeResponse: any, userContext: any) => Promise<UserInfo>
  getClientId: (userContext: any) => string
  getRedirectURI?: (userContext: any) => string // if undefined, the redirect_uri is set on the frontend.
}

export interface TypeProvider {
  id: string
  get: (
    redirectURI: string | undefined,
    authCodeFromRequest: string | undefined,
    userContext: any
  ) => TypeProviderGetResponse
  isDefault?: boolean // if not present, we treat it as false
}