// src/recipe/thirdpartypasswordless/types.ts
export type DeviceType = DeviceTypeOriginal

export type User = (
  | {
    // passwordless user properties
    email?: string
    phoneNumber?: string
  }
  | {
    // third party user properties
    email: string
    thirdParty: {
      id: string
      userId: string
    }
  }
) & {
  id: string
  timeJoined: number
}
// ...
export interface RecipeInterface {
  getUserById(input: { userId: string; userContext: any }): Promise<User | undefined>

  getUsersByEmail(input: { email: string; userContext: any }): Promise<User[]>

  getUserByPhoneNumber: (input: { phoneNumber: string; userContext: any }) => Promise<User | undefined>

  getUserByThirdPartyInfo(input: {
    thirdPartyId: string
    thirdPartyUserId: string
    userContext: any
  }): Promise<User | undefined>

  thirdPartySignInUp(input: {
    thirdPartyId: string
    thirdPartyUserId: string
    email: string
    userContext: any
  }): Promise<{ status: 'OK'; createdNewUser: boolean; user: User }>

  createCode: (
    input: (
      | {
        email: string
      }
      | {
        phoneNumber: string
      }
    ) & { userInputCode?: string; userContext: any }
  ) => Promise<{
    status: 'OK'
    preAuthSessionId: string
    codeId: string
    deviceId: string
    userInputCode: string
    linkCode: string
    codeLifetime: number
    timeCreated: number
  }>

  createNewCodeForDevice: (input: {
    deviceId: string
    userInputCode?: string
    userContext: any
  }) => Promise<
        | {
          status: 'OK'
          preAuthSessionId: string
          codeId: string
          deviceId: string
          userInputCode: string
          linkCode: string
          codeLifetime: number
          timeCreated: number
        }
        | { status: 'RESTART_FLOW_ERROR' | 'USER_INPUT_CODE_ALREADY_USED_ERROR' }
    >

  consumeCode: (
    input:
    | {
      userInputCode: string
      deviceId: string
      preAuthSessionId: string
      userContext: any
    }
    | {
      linkCode: string
      preAuthSessionId: string
      userContext: any
    }
  ) => Promise<
        | {
          status: 'OK'
          createdNewUser: boolean
          user: User
        }
        | {
          status: 'INCORRECT_USER_INPUT_CODE_ERROR' | 'EXPIRED_USER_INPUT_CODE_ERROR'
          failedCodeInputAttemptCount: number
          maximumCodeInputAttempts: number
        }
        | { status: 'RESTART_FLOW_ERROR' }
    >

  updatePasswordlessUser: (input: {
    userId: string
    email?: string | null
    phoneNumber?: string | null
    userContext: any
  }) => Promise<{
    status: 'OK' | 'UNKNOWN_USER_ID_ERROR' | 'EMAIL_ALREADY_EXISTS_ERROR' | 'PHONE_NUMBER_ALREADY_EXISTS_ERROR'
  }>

  revokeAllCodes: (
    input:
    | {
      email: string
      userContext: any
    }
    | {
      phoneNumber: string
      userContext: any
    }
  ) => Promise<{
    status: 'OK'
  }>

  revokeCode: (input: {
    codeId: string
    userContext: any
  }) => Promise<{
    status: 'OK'
  }>

  listCodesByEmail: (input: { email: string; userContext: any }) => Promise<DeviceType[]>

  listCodesByPhoneNumber: (input: { phoneNumber: string; userContext: any }) => Promise<DeviceType[]>

  listCodesByDeviceId: (input: { deviceId: string; userContext: any }) => Promise<DeviceType | undefined>

  listCodesByPreAuthSessionId: (input: {
    preAuthSessionId: string
    userContext: any
  }) => Promise<DeviceType | undefined>
}

export type PasswordlessAPIOptions = PasswordlessAPIOptionsOriginal

export type ThirdPartyAPIOptions = ThirdPartyAPIOptionsOriginal
export interface APIInterface {
  authorisationUrlGET:
  | undefined
  | ((input: {
    provider: TypeProvider
    options: ThirdPartyAPIOptions
    userContext: any
  }) => Promise<
              | {
                status: 'OK'
                url: string
              }
              | GeneralErrorResponse
          >)

  thirdPartySignInUpPOST:
  | undefined
  | ((input: {
    provider: TypeProvider
    code: string
    redirectURI: string
    authCodeResponse?: any
    clientId?: string
    options: ThirdPartyAPIOptions
    userContext: any
  }) => Promise<
              | {
                status: 'OK'
                createdNewUser: boolean
                user: User
                session: SessionContainerInterface
                authCodeResponse: any
              }
              | GeneralErrorResponse
              | {
                status: 'NO_EMAIL_GIVEN_BY_PROVIDER'
              }
          >)

  appleRedirectHandlerPOST:
  | undefined
  | ((input: { code: string; state: string; options: ThirdPartyAPIOptions; userContext: any }) => Promise<void>)

  createCodePOST:
  | undefined
  | ((
    input: ({ email: string } | { phoneNumber: string }) & {
      options: PasswordlessAPIOptions
      userContext: any
    }
  ) => Promise<
              | {
                status: 'OK'
                deviceId: string
                preAuthSessionId: string
                flowType: 'USER_INPUT_CODE' | 'MAGIC_LINK' | 'USER_INPUT_CODE_AND_MAGIC_LINK'
              }
              | GeneralErrorResponse
          >)

  resendCodePOST:
  | undefined
  | ((
    input: { deviceId: string; preAuthSessionId: string } & {
      options: PasswordlessAPIOptions
      userContext: any
    }
  ) => Promise<GeneralErrorResponse | { status: 'RESTART_FLOW_ERROR' | 'OK' }>)

  consumeCodePOST:
  | undefined
  | ((
    input: (
      | {
        userInputCode: string
        deviceId: string
        preAuthSessionId: string
      }
      | {
        linkCode: string
        preAuthSessionId: string
      }
    ) & {
      options: PasswordlessAPIOptions
      userContext: any
    }
  ) => Promise<
              | {
                status: 'OK'
                createdNewUser: boolean
                user: User
                session: SessionContainerInterface
              }
              | {
                status: 'INCORRECT_USER_INPUT_CODE_ERROR' | 'EXPIRED_USER_INPUT_CODE_ERROR'
                failedCodeInputAttemptCount: number
                maximumCodeInputAttempts: number
              }
              | GeneralErrorResponse
              | { status: 'RESTART_FLOW_ERROR' }
          >)

  passwordlessUserEmailExistsGET:
  | undefined
  | ((input: {
    email: string
    options: PasswordlessAPIOptions
    userContext: any
  }) => Promise<
              | {
                status: 'OK'
                exists: boolean
              }
              | GeneralErrorResponse
          >)

  passwordlessUserPhoneNumberExistsGET:
  | undefined
  | ((input: {
    phoneNumber: string
    options: PasswordlessAPIOptions
    userContext: any
  }) => Promise<
              | {
                status: 'OK'
                exists: boolean
              }
              | GeneralErrorResponse
          >)
}

export type TypeThirdPartyPasswordlessEmailDeliveryInput = TypePasswordlessEmailDeliveryInput